+++ HISTORY.md
@@ -1,3 +1,7 @@
+unversioned: tt-bot-dev changes
+=================================
+  * configurable options for each request
+
 unreleased
 ==========